backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fastapi>=0.135.2",
    "lxml>=6.0.2",
    "mypy>=1.19.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "requests>=2.33.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "types-psycopg2>=2.9.21.20260223",
    "types-requests>=2.33.0.20260327",
]
